SMT_attendance/requirements.txt
alembic==1.4.1
asgiref==3.2.4
certifi==2019.11.28
chardet==3.0.4
Click==7.0
cmake==3.16.3
Django==3.0.4
dlib==19.19.0
face-recognition==1.3.0
face-recognition-models==0.3.0
Flask==1.1.1
Flask-Migrate==2.5.2
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.1
idna==2.9
image==1.5.28
itsdangerous==1.1.0
Jinja2==2.11.1
Mako==1.1.2
MarkupSafe==1.1.1
migrate==0.3.8
numpy==1.18.1
Pillow==7.0.0
psycopg2-binary==2.8.4
python-dateutil==2.8.1
python-editor==1.0.4
pytz==2019.3
requests==2.23.0
six==1.14.0
SQLAlchemy==1.3.13
sqlparse==0.3.1
urllib3==1.25.8
Werkzeug==0.16.1
